+++ example/app/tables.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python
 # coding: utf-8
 
+from datetime import date
 from table.columns import Column, LinkColumn, DatetimeColumn, Link
+from table.columns.calendarcolumn import MonthsColumn, WeeksColumn, DaysColumn
 from table.utils import A
 from table import Table
 from models import Person
@@ -14,7 +16,6 @@ class PersonTable(Table):
 
     class Meta:
         model = Person
-        ext_button_template = "button.html"
         # disable_search = True
         # disable_info = True
         # disable_length_menu = True
@@ -23,8 +24,7 @@ class PersonTable(Table):
 class ScheduleTable(Table):
     id = Column(field='id', header=u'#', header_attrs={'width': '5%'})
     name = Column(field='name', header=u'NAME', header_attrs={'width': '15%'})
-    
-    
-    class Meta:
-        model = Person
+    # days = DaysColumn(field='calendar', start_date=date(2013, 12, 18), end_date=date(2014, 1, 1))
+    # weeks = WeeksColumn(field='calendar', start_date=date(2013, 12, 18), end_date=date(2014, 1, 1))
+    month = MonthsColumn(field='calendar', start_date=date(2013, 12, 18), end_date=date(2014, 5, 1))