+++ example/settings.py
@@ -7,7 +7,7 @@ import sys
 sys.path.append('..')
 
 DEBUG = True
-TEMPLATE_DEBUG = DEBUG
+TEMPLATE_DEBUG = False
 
 ADMINS = (
     # ('Your Name', 'your_email@example.com'),