+++ table/columns/__init__.py
@@ -4,3 +4,5 @@
 from .base import Column, BoundColumn
 from .linkcolumn import LinkColumn, Link
 from .datetimecolumn import DatetimeColumn
+from .calendarcolumn import MonthsColumn, WeeksColumn, DaysColumn
+from .sequencecolumn import SequenceColumn