+++ table/tables.py
@@ -1,13 +1,14 @@
 #!/usr/bin/env python
 # coding: utf-8
 
+import copy
+import traceback
 from django.db.models.query import QuerySet
 from django.utils.safestring import mark_safe
 from django.utils.datastructures import SortedDict
-from columns import Column, BoundColumn
-from addon import *
-import copy
-import traceback
+from columns import Column, BoundColumn, SequenceColumn
+from addon import (TableSearchBox, TableInfoLabel, TablePagination,
+                   TableLengthMenu, TableExtButton)
 
 
 class BaseTable(object):
@@ -29,19 +30,15 @@ class BaseTable(object):
     @property
     def rows(self):
         rows = []
-        try:
-            for obj in self.data:
-                # Binding object to each column of each row, so that
-                # data structure for each row is organized like this:
-                # { boundcol0: td, boundcol1: td, boundcol2: td }
-                row = SortedDict()
-                columns = [BoundColumn(obj, col) for col in self.columns if col.space]
-                for col in columns:
-                    row[col] = col.html
-                rows.append(row)
-        except Exception, e:
-            print e
-            print traceback.format_exc()
+        for obj in self.data:
+            # Binding object to each column of each row, so that
+            # data structure for each row is organized like this:
+            # { boundcol0: td, boundcol1: td, boundcol2: td }
+            row = SortedDict()
+            columns = [BoundColumn(obj, col) for col in self.columns if col.space]
+            for col in columns:
+                row[col] = col.html
+            rows.append(row)
         return rows
 
     @property
@@ -49,6 +46,7 @@ class BaseTable(object):
         """
         [ [header1], [header3, header4] ]
         """
+        # TO BE FIX: refactor
         header_rows = []
         headers = [col.header for col in self.columns]
         for header in headers:
@@ -59,40 +57,57 @@ class BaseTable(object):
 
 class TableAddons(object):
     def __init__(self, table):
-        self.search_box = TableSearchBox(table.opts.search_placeholder,
-                                         table.opts.disable_search)
-        self.info_label = TableInfoLabel(table.opts.info_format,
-                                         table.opts.disable_info)
-        self.pagination = TablePagination(table.opts.pagination_first,
-                                          table.opts.pagination_last,
-                                          table.opts.pagination_prev,
-                                          table.opts.pagination_next,
-                                          table.opts.disable_pagination)
-        self.length_menu = TableLengthMenu(table.opts.disable_length_menu)
-        self.ext_button = TableExtButton(table.opts.ext_button_template,
-                                         table.opts.ext_button_context)
+        options = table.opts
+        self.length_menu = TableLengthMenu(visible=options.length_menu)
+        self.info_label = TableInfoLabel(format=options.info_format,
+                                         visible=options.info)
+        self.search_box = TableSearchBox(placeholder=options.search_placeholder,
+                                         visible=options.search)
+        self.ext_button = TableExtButton(template=options.ext_button_template,
+                                         context=options.ext_button_context,
+                                         visible=options.ext_button)
+        self.pagination = TablePagination(first=options.pagination_first,
+                                          last=options.pagination_last,
+                                          prev=options.pagination_prev,
+                                          next=options.pagination_next,
+                                          visible=options.pagination)
 
     def render_dom(self):
-        dom = "<'row'"
-        dom += ''.join([self.ext_button.dom, self.search_box.dom])
-        dom += "r>t<'row'"
-        dom += ''.join([self.info_label.dom, self.pagination.dom, self.length_menu.dom])
-        dom += ">"
+        dom = ''
+        if self.search_box.visible or self.ext_button.visible:
+            dom += "<'row'" + ''.join([self.ext_button.dom, self.search_box.dom]) + ">"
+        dom += "rt"
+        if self.info_label.visible or self.pagination.visible or self.length_menu.visible:
+            dom += "<'row'" + ''.join([self.info_label.dom, self.pagination.dom, self.length_menu.dom]) + ">"
         return mark_safe(dom)
 
 class TableOptions(object):
-    def __init__(self, clsname, options=None):
+    def __init__(self, options=None):
         self.model = getattr(options, 'model', None)
 
-        # take class name in lowcase as default id of <table>
-        self.id = getattr(options, 'id', clsname.lower())
+        # id attribute of <table> tag
+        self.id = getattr(options, 'id', None)
 
         # build attributes for <table> tag, use bootstrap
         # css class "table table-boarded" as default style
         attrs = getattr(options, 'attrs', {})
-        attrs['class'] = 'table table-bordered' + attrs.get('class', '')
-        self.attrs = mark_safe(' '.join(['%s="%s"' % (attr_name, attr) for attr_name, attr in attrs.items()]))
+        attrs['class'] = 'table table-bordered ' + attrs.get('class', '')
+        self.attrs = mark_safe(' '.join(['%s="%s"' % (attr_name, attr)
+                                         for attr_name, attr in attrs.items()]))
+        # build attributes for <thead> and <tbody>
+        thead_attrs = getattr(options, 'thead_attrs', {})
+        self.thead_attrs = mark_safe(' '.join(['%s="%s"' % (attr_name, attr)
+                                         for attr_name, attr in thead_attrs.items()]))
+        tbody_attrs = getattr(options, 'tbody_attrs', {})
+        self.tbody_attrs = mark_safe(' '.join(['%s="%s"' % (attr_name, attr)
+                                         for attr_name, attr in tbody_attrs.items()]))
 
+        # scrolling option
+        self.scrollable = getattr(options, 'scrollable', False)
+        self.scrollinner = getattr(options, 'scrollinner', "150%")
+        self.fixed_columns = getattr(options, 'fixed_columns', None)
+        self.fixed_columns_width = getattr(options, 'fixed_columns_width', None)
+        
         # inspect sorting option
         self.sort = []
         for column, order in getattr(options, 'sort', []):
@@ -105,21 +120,24 @@ class TableOptions(object):
             self.sort.append((column, order))
 
         # options for table add-on
-        self.disable_search = getattr(options, 'disable_search', False)
+        self.search = getattr(options, 'search', True)
         self.search_placeholder = getattr(options, 'search_placeholder', 'Search')
 
-        self.disable_info = getattr(options, 'disable_info', False)
+        self.info = getattr(options, 'info', True)
         self.info_format = getattr(options, 'info_format', 'Total _TOTAL_')
 
-        self.disable_pagination = getattr(options, 'disable_pagination', False)
+        self.pagination = getattr(options, 'pagination', True)
         self.pagination_first = getattr(options, 'pagination_first', 'First')
         self.pagination_last = getattr(options, 'pagination_last', 'Last')
         self.pagination_prev = getattr(options, 'pagination_prev', 'Prev')
         self.pagination_next = getattr(options, 'pagination_next', 'Next')
 
-        self.disable_length_menu = getattr(options, 'disable_length_menu', False)
+        self.length_menu = getattr(options, 'length_menu', True)
+
+        self.ext_button = getattr(options, 'ext_button', False)
         self.ext_button_template = getattr(options, 'ext_button_template', None)
         self.ext_button_context = getattr(options, 'ext_button_context', None)
+
         self.zero_records = getattr(options, 'zero_records', u'No records')
 
 class TableMetaClass(type):
@@ -127,20 +145,31 @@ class TableMetaClass(type):
     """
 
     def __new__(cls, name, bases, attrs):
-        columns, meta = [], None
+        opts = TableOptions(attrs.get('Meta', None))
+        if not opts.id:
+            opts.id = name.lower()
+        attrs['opts'] = opts
 
-        # extract declared columns and meta
+        # extract declared columns
+        columns = []
         for attr_name, attr in attrs.items():
-            if isinstance(attr, Column):
+            if isinstance(attr, SequenceColumn):
+                columns.extend(attr.extract())
+            elif isinstance(attr, Column):
                 columns.append(attr)
-            else:
-                meta = attr
         columns.sort(key=lambda x: x.instance_order)
-        attrs['base_columns'] = columns
-        attrs['opts'] = TableOptions(name, meta)
+
+        # If this class is subclassing other tables, add their fields as
+        # well. Note that we loop over the bases in reverse - this is
+        # necessary to preserve the correct order of columns.
+        parent_columns = []
+        for base in bases[::-1]:
+            if hasattr(base, "base_columns"):
+                parent_columns = base.base_columns + parent_columns
+
+        attrs['base_columns'] = parent_columns + columns
 
         return super(TableMetaClass, cls).__new__(cls, name, bases, attrs)
 
 
 Table = TableMetaClass('Table', (BaseTable,), {})
-