+++ table/tests/__init__.py
@@ -1,0 +1,3 @@
+from test_columns import *
+from test_tables import *
+